An important update
updated,

1.imported more files,
2. login
3.logout
4.declared odooclient and odoo session

important note! right now min sdk is 2.7, where the null safety feature is still not activated.

upon finish migrating, if wish to update SDK to 2.10, we will have to change a lot of things to adjust to the null safety feature.

for now, keep with 2.7sdk

This is an update to prepare for pulling data from odoo using API call